requirements.txt
cohere>=5.0.0
python-dotenv>=1.0.0
chromadb>=0.4.0
streamlit>=1.28.0
tqdm>=4.65.0
